refactor: update cli to use unified scope api
Consolidate all member management functions from org.ts into scopes.ts
using scopeMembersContract. Delete cli org.ts domain file. Update cli
commands and tests to use /api/scope/* endpoints.
