+++ turbo/apps/cli/src/commands/scope/__tests__/leave.test.ts
@@ -39,7 +39,7 @@ describe("org leave command", () => {
 
   it("should leave org and show success", async () => {
     server.use(
-      http.post("http://localhost:3000/api/org/leave", () => {
+      http.post("http://localhost:3000/api/scope/leave", () => {
         return HttpResponse.json({
           message: "Left organization",
         });
@@ -55,7 +55,7 @@ describe("org leave command", () => {
 
   it("should handle admin-cannot-leave error", async () => {
     server.use(
-      http.post("http://localhost:3000/api/org/leave", () => {
+      http.post("http://localhost:3000/api/scope/leave", () => {
         return HttpResponse.json(
           {
             error: {
