+++ turbo/apps/cli/src/commands/scope/__tests__/members.test.ts
@@ -30,7 +30,7 @@ describe("org status command", () => {
 
   it("should display org name, role, and member list", async () => {
     server.use(
-      http.get("http://localhost:3000/api/org/status", () => {
+      http.get("http://localhost:3000/api/scope/members", () => {
         return HttpResponse.json({
           slug: "my-team",
           role: "admin",
@@ -65,7 +65,7 @@ describe("org status command", () => {
 
   it("should show helpful error when no org scope active", async () => {
     server.use(
-      http.get("http://localhost:3000/api/org/status", () => {
+      http.get("http://localhost:3000/api/scope/members", () => {
         return HttpResponse.json(
           {
             error: {
