+++ turbo/apps/cli/src/commands/scope/remove.ts
@@ -1,6 +1,6 @@
 import { Command } from "commander";
 import chalk from "chalk";
-import { removeMember } from "../../lib/api";
+import { removeScopeMember } from "../../lib/api";
 
 export const removeCommand = new Command()
   .name("remove")
@@ -8,7 +8,7 @@ export const removeCommand = new Command()
   .argument("<email>", "Email address of the member to remove")
   .action(async (email: string) => {
     try {
-      await removeMember(email);
+      await removeScopeMember(email);
       console.log(chalk.green(`✓ Removed ${email} from scope`));
     } catch (error) {
       if (error instanceof Error) {
